% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CFMortality.R
\name{predictCFMortality}
\alias{predictCFMortality}
\title{1- and 2- year Mortality Prediction Models in Cystic Fibrosis (CF)}
\source{
\url{https://erj.ersjournals.com/content/early/2019/05/08/13993003.00224-2019}
}
\usage{
predictCFMortality(age, male, fvc, fev1, fev1LastYear, bcepacia,
  underweight, nHosp, pancreaicInsufficient, CFRelatedDiabetes,
  ageAtDiagnosis)
}
\arguments{
\item{age}{A number for age}

\item{male}{A binary variable with 0 for females and 1 for males}

\item{fvc}{A number for percentage of FVC in current year}

\item{fev1}{A number for percentage of fev11 in current year}

\item{fev1LastYear}{A number for percentage of fev11 in preceeding year}

\item{bcepacia}{A binary with 0 for no B. cepacia complex and 1 for B. cepacia complex}

\item{underweight}{A binary with 0 for not being underweight (BMI<18.5) and 1 otherwise}

\item{nHosp}{number of hospancreaicInsufficienttalizations in preceding year}

\item{pancreaicInsufficient}{A binary taking 1 for pancreaic insufficient status and 0 otherwise}

\item{CFRelatedDiabetes}{A binary variable for CF related diabetes}

\item{ageAtDiagnosis}{A number for age at CF diagnosis in years}
}
\description{
1- and 2- year Mortality Prediction Models in Cystic Fibrosis (CF)
}
\examples{
predictCFMortality (age = 16, male = 0, fvc = 66.7, fev1 = 47.4, fev1LastYear = 80.5, bcepacia = 0, underweight = 0, nHosp = 0, pancreaicInsufficient = 1, CFRelatedDiabetes = 0, ageAtDiagnosis = 0.9)
predictCFMortality (age = 40.4, male = 1, fvc = 25.7, fev1 = 19.2, fev1LastYear = 20, bcepacia = 1, underweight = 1, nHosp = 6, pancreaicInsufficient = 0, CFRelatedDiabetes = 0, ageAtDiagnosis = 27.2)
}
