% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpmcf.R
\name{predictCFMortality}
\alias{predictCFMortality}
\title{Predicting 1- and 2-year risk of death for patients with cystic fibrosis (CF)}
\source{
\url{https://erj.ersjournals.com/content/early/2019/05/08/13993003.00224-2019}
}
\usage{
predictCFMortality(age, gender, fvc, fev, fevp, bcepacia, uw, h, pi, cfrd,
  aged)
}
\arguments{
\item{age}{A number for age, first argument}

\item{gender}{A binary variable with 0 for females and 1 for males, second arugment}

\item{fvc}{A number for percentage of FVC in current year, 3rd argument}

\item{fev}{A number for percentage of FEV1 in current year, 4th argument}

\item{fevp}{A number for percentage of FEV1 in preceeding year, 5th argument}

\item{bcepacia}{A binary with 0 for no B. cepacia complex and 1 for B. cepacia complex, sixth argument}

\item{uw}{A binary with 0 for not being under weight and 1 otherwise, 7th argument}

\item{h}{A number for number of hospitalizations in preceding year, 8th argument}

\item{pi}{A binary taking 1 for pancreaic insufficient status and 0 otherwise, 9th argument}

\item{cfrd}{A binary taking 0 for not having cf related disease and 1 for having cf related disease, 10th argument}

\item{aged}{A number for age at CF diagnosis in years, 11th argument}
}
\value{
The code
}
\description{
Predicting 1- and 2-year risk of death for patients with cystic fibrosis (CF)
}
\examples{
predictCFMortality(16, 0, 66.7, 47.4, 80.5, 0, 0, 0, 1, ,0 , 0.9)
}
